Move typedefs around
